chore: split up TextLine into its own file
